valuation-service/requirements.txt
fastapi==0.119.0
uvicorn[standard]==0.37.0
scikit-learn==1.8.0
numpy==2.3.3
pandas==2.3.3
mysql-connector-python==9.6.0
python-dotenv==1.2.2
geohash2==1.1
haversine==2.9.0
joblib==1.5.0
scipy==1.17.1
pydantic==2.11.2
